pet-py/pyproject.toml
[project]
name = "pet-py"
version = "0.1.0"
description = "Add your description here"
dependencies = [
    "more-itertools>=10.3.0",
    "deepbet>=0.0.2",
    "pandas[performance,computation,compression]>=2.2.2",
    "numexpr>=2.10.1",
    "nibabel>=5.2.1",
    "nilearn>=0.10.4",
    "nipype>=1.8.6",
    "itk-elastix>=0.20.0",
    "fslpy>=3.19.1",
    "pydicom>=2.4.4",
    "pandas-stubs>=2.2.2.240603",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "pytest>=8.2.2",
    "ipython>=8.26.0",
    "pyjanitor>=0.27.0",
    "mypy>=1.10.1",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/pet_py"]

[tool.rye.scripts]
lint = { chain = ["lint:ruff", "lint:black"] }
"lint:black" = "black --check src"
"lint:ruff" = "ruff check src"



[tool.mypy]
disable_error_code = ["import-untyped", "attr-defined"]
